Split up Reference
