+++ src/Components/Reference.php
@@ -6,10 +6,6 @@ namespace PhpMyAdmin\SqlParser\Components;
 
 use PhpMyAdmin\SqlParser\Component;
 use PhpMyAdmin\SqlParser\Context;
-use PhpMyAdmin\SqlParser\Parseable;
-use PhpMyAdmin\SqlParser\Parser;
-use PhpMyAdmin\SqlParser\TokensList;
-use PhpMyAdmin\SqlParser\TokenType;
 
 use function implode;
 use function trim;
@@ -17,26 +13,8 @@ use function trim;
 /**
  * `REFERENCES` keyword parser.
  */
-final class Reference implements Component, Parseable
+final class Reference implements Component
 {
-    /**
-     * All references options.
-     */
-    private const REFERENCES_OPTIONS = [
-        'MATCH' => [
-            1,
-            'var',
-        ],
-        'ON DELETE' => [
-            2,
-            'var',
-        ],
-        'ON UPDATE' => [
-            3,
-            'var',
-        ],
-    ];
-
     /**
      * The referenced table.
      *
@@ -68,69 +46,6 @@ final class Reference implements Component, Parseable
         $this->table = $table;
         $this->columns = $columns;
         $this->options = $options;
-    }
-
-    /**
-     * @param Parser               $parser  the parser that serves as context
-     * @param TokensList           $list    the list of tokens that are being parsed
-     * @param array<string, mixed> $options parameters for parsing
-     */
-    public static function parse(Parser $parser, TokensList $list, array $options = []): Reference
-    {
-        $ret = new static();
-
-        /**
-         * The state of the parser.
-         *
-         * Below are the states of the parser.
-         *
-         *      0 ----------------------[ table ]---------------------> 1
-         *
-         *      1 ---------------------[ columns ]--------------------> 2
-         *
-         *      2 ---------------------[ options ]--------------------> (END)
-         */
-        $state = 0;
-
-        for (; $list->idx < $list->count; ++$list->idx) {
-            /**
-             * Token parsed at this moment.
-             */
-            $token = $list->tokens[$list->idx];
-
-            // End of statement.
-            if ($token->type === TokenType::Delimiter) {
-                break;
-            }
-
-            // Skipping whitespaces and comments.
-            if (($token->type === TokenType::Whitespace) || ($token->type === TokenType::Comment)) {
-                continue;
-            }
-
-            if ($state === 0) {
-                $ret->table = Expression::parse(
-                    $parser,
-                    $list,
-                    [
-                        'parseField' => 'table',
-                        'breakOnAlias' => true,
-                    ],
-                );
-                $state = 1;
-            } elseif ($state === 1) {
-                $ret->columns = ArrayObj::parse($parser, $list)->values;
-                $state = 2;
-            } elseif ($state === 2) {
-                $ret->options = OptionsArray::parse($parser, $list, self::REFERENCES_OPTIONS);
-                ++$list->idx;
-                break;
-            }
-        }
-
-        --$list->idx;
-
-        return $ret;
     }
 
     public function build(): string
